Added 'rebin' option for linout plots.
This allows lineouts to be plotted with better statistics per point
without introducing correlations from one point to the next as is
the case with the 'smooth' arguments to the lineout plotting methods
in camcontrol.py.

Also added support for plotly-based plots embedded in the Jupyter
notebook. Plotly mode is enabled by setting config.plotting_mode
to 'notebook'.