Bring this package up to date with camera driver modification.
With the -w N command line parameter, the driver writes the sum of frames
incrementally every N frames, to a file with the additional suffix .part. Each
time it does this it also applies pixel value windowing to the incremental sum
of frames (in addition to frame-by-frame filtering), if cluster rejection mode
is selected.

This removes the need to break up long exposure sequences: we save disk space
by offloading the frame-summing that was previously done in camcontrol.py. With
a minimal additional change we could now have camcontrol start exposure
sequences without blocking.

Note that the interfaces of camcontrol.RunSet and the utility functions in nbinit.py have changed.

Additional changes:
	-Allow accessing sum-of-frames data of an ongoing exposure sequence.
	-Add a plot_histogram method to camcontrol.Frame
	-Have camcontrol.plot_lineout call plt.show() by default