Fix camera's aborted readouts in non-blocking mode.
The problem seems to have been that paramiko's ssh connection was timing
out after about 30 minutes. Worked around it by launching main_mt9m001
in screen.

One result of this change is that stdout from main_mt9m001 can no longer
be captured client-side. Instead it's written to a logfile on the
Beaglebone.